requirements.txt
fastapi
uvicorn
requests
websocket-client
py-clob-client
python-dotenv
pydantic